moved custom snack bar to a separate widget
