+++ aptiche/lib/utils/theme.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 
 //colour used for buttons and other widgets
 const Color kPrimaryColor = Color(0xff3385be);
@@ -75,11 +74,4 @@ ThemeData appTheme() {
       ),
     ),
   );
-}
-
-void customSnackBar(String title, String subject) {
-  Get.snackbar<dynamic>(title, subject,
-      backgroundColor: kSnackColour.withOpacity(0.5),
-      snackStyle: SnackStyle.FLOATING,
-      colorText: Colors.black87);
 }
