+++ aptiche/lib/widgets/snackbar.dart
@@ -1,0 +1,12 @@
+import 'package:aptiche/utils/theme.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+
+class CustomLoaders {
+  void customSnackBar(String title, String subject) {
+    Get.snackbar<dynamic>(title, subject,
+        backgroundColor: kSnackColour.withOpacity(0.5),
+        snackStyle: SnackStyle.FLOATING,
+        colorText: Colors.black87);
+  }
+}
